// web/src/lib/api/model/fieldKey.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Silo
 * OpenAPI spec version: 0.1.0
 */

export type FieldKey = (typeof FieldKey)[keyof typeof FieldKey]

export const FieldKey = {
  title: "title",
  sort_title: "sort_title",
  description_short: "description_short",
  description_full: "description_full",
  first_release_date: "first_release_date",
  wrapper: "wrapper",
  game_type: "game_type",
  developers: "developers",
  publishers: "publishers",
  series: "series",
  genres: "genres",
  themes: "themes",
  tags: "tags",
  engines: "engines",
  modes: "modes",
  platforms: "platforms",
  release_dates: "release_dates",
  ratings: "ratings",
  sys_requirements: "sys_requirements",
  alt_names: "alt_names",
  videos: "videos",
  cover: "cover",
  background: "background",
  landscape: "landscape",
  logo: "logo",
  icon: "icon",
  screenshot: "screenshot",
  video_thumbnail: "video_thumbnail",
} as const